+++ lib/STM_sequential.mli
@@ -4,10 +4,13 @@ module Make : functor (Spec : STM.Spec) ->
   sig
     val cmds_ok : Spec.state -> Spec.cmd list -> bool
     (** A precondition checker (stops early, thanks to short-circuit Boolean evaluation).
-        Accepts the initial state and the command sequence as parameters.  *)
+        Accepts the initial state and the command sequence as parameters.
+        [cmds_ok] catches and ignores exceptions arising from {!next_state}.  *)
 
     val arb_cmds : Spec.state -> Spec.cmd list QCheck.arbitrary
-    (** A generator of {!Spec.cmd} sequences. Accepts the initial state as a parameter. *)
+    (** A generator of {!Spec.cmd} sequences. Accepts the initial state as a parameter.
+        [arb_cmds] catches and ignores generation-time exceptions arising from
+        {!Spec.next_state}. *)
 
     val agree_prop : Spec.cmd list -> bool
     (** The agreement property: the command sequence [cs] yields the same observations